+++ sdk/release_notes.md
@@ -4,7 +4,11 @@
 - My change description (#PR/#issue)
 -->
 
-### Microsoft.Azure.Functions.Worker.Sdk 1.17.2 (meta package)
+### Microsoft.Azure.Functions.Worker.Sdk 1.17.3
 
-- Revert changes from #1946
+- Updating to use `Microsoft.NET.Sdk.Functions.Generators` 1.2.2 (#2247)
+
+### Microsoft.Azure.Functions.Worker.Sdk.Generators 1.2.2
+
+- Introduces handling for `HttpResultAttribute`.